requirements.txt
streamlit
plotly
pandas
firebase-admin
streamlit-autorefresh
pyrebase4


